Split buffer plugin into modules.
